Fix CLI parsing and validation issues
- Fixed CLI argument conflicts in src/config/cli.rs (removed duplicate short options)
- Removed dead code (unused to_config function) that was causing 73+ compilation errors
- Fixed clippy lints: useless_conversion, unnecessary_cast, collapsible_if, doc formatting
- Fixed all formatting issues with rustfmt
- All tests passing: 253 library tests + 15 ad-hoc CLI tests
- Module compiles with zero errors and zero warnings

+++ src/network/conntrack.rs
@@ -425,8 +425,8 @@ mod tests {
 
     #[test]
     fn test_protocol_to_libc() {
-        assert_eq!(Protocol::Udp.to_libc(), libc::IPPROTO_UDP as i32);
-        assert_eq!(Protocol::Tcp.to_libc(), libc::IPPROTO_TCP as i32);
+        assert_eq!(Protocol::Udp.to_libc(), libc::IPPROTO_UDP);
+        assert_eq!(Protocol::Tcp.to_libc(), libc::IPPROTO_TCP);
     }
 
     #[test]
